ml/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.0
supabase==2.4.3
python-dotenv==1.0.1
geopy==2.4.1
requests==2.31.0
httpx==0.27.0
pydantic==2.7.1
